repoResponseList 분리
